Restaurant Website
This restaurant website was built entirely with HTML and CSS. It has a lovely homepage with an interactive menu, a contact page with a Google Map attached, a blog page with a prepared post, and an about page. The website is also responsive, which means it will appear fantastic on any device. Positioning items, making the website responsive, and designing the page using CSS are some complex topics that were employed in the creation of this website.